fixed : issue at cloudinary.js

+++ src/utils/cloudinary.js
@@ -1,7 +1,7 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
-import dotenv from "dotenv"
-dotenv.config()
+import dotenv from "dotenv";
+dotenv.config();
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -18,7 +18,9 @@ const uploadOnCloudinary = async (localFilePath) => {
     });
     // file has been uploaded successfull
     // console.log("file is uploaded on cloudinary ", response.url);
-    fs.unlinkSync(localFilePath);
+    if (response) {
+      fs.unlinkSync(localFilePath);
+    }
     return response;
   } catch (error) {
     fs.unlinkSync(localFilePath); // r.emove the locally saved temporary file as the upload operation got failed
